feat(spo): rename spo web alert to spo web rule (alert as alias)
Per @Adam-it's review on #7221, inverts the alias direction:
- spo web rule list/remove are the canonical commands
- spo web alert list/remove remain available as aliases for
  backward compatibility, which makes the eventual removal of the
  alert names a single-line change.

Renames the four command files (web-alert-{list,remove}.{ts,spec.ts}
-> web-rule-{list,remove}.{ts,spec.ts}), updates the command name
constants in src/m365/spo/commands.ts, switches the alias() return
to the legacy WEB_ALERT_* constants, renames the two docs files and
adds the ## Alias section pointing at the legacy command, retargets
the docs sidebar entries, and updates the sample script + release
note to reference the canonical web-rule name.

Verification: clean build + lint + 35 focused specs pass. Full
npm test was blocked locally on the Node 24 version gate (this
shell is on Node 25) and could not fetch Node 24 from sandbox;
CI is the authority for the full matrix.
