+++ docs/docs/about/release-notes.mdx
@@ -251,8 +251,8 @@ sidebar_position: 4
 
 **SharePoint:**
 
-- [spo web alert list](../cmd/spo/web/web-alert-list.mdx) - lists all SharePoint list alerts [#6861](https://github.com/pnp/cli-microsoft365/issues/6861)
-- [spo web alert remove](../cmd/spo/web/web-alert-remove.mdx) - removes an alert from a SharePoint list [#6863](https://github.com/pnp/cli-microsoft365/issues/6863)
+- [spo web rule list](../cmd/spo/web/web-rule-list.mdx) - lists all SharePoint list alerts [#6861](https://github.com/pnp/cli-microsoft365/issues/6861)
+- [spo web rule remove](../cmd/spo/web/web-rule-remove.mdx) - removes an alert from a SharePoint list [#6863](https://github.com/pnp/cli-microsoft365/issues/6863)
 
 ### Changes
