+++ docs/docs/sample-scripts/spo/list-tenant-alert-usage/index.mdx
@@ -77,7 +77,7 @@ This script assumes you have permissions to all sites in the tenant. We recommen
 
     try {
       # Get deprecated alerts from specific site
-      $alerts = m365 spo web alert list --webUrl $SiteUrl --output json --query "[?!(Properties[].Key && contains(Properties[].Key, 'ruletitle'))]" | ConvertFrom-Json
+      $alerts = m365 spo web rule list --webUrl $SiteUrl --output json --query "[?!(Properties[].Key && contains(Properties[].Key, 'ruletitle'))]" | ConvertFrom-Json
 
       if ($alerts.Count -gt 0) {
         Write-Host "`tFound $($alerts.Count) alert(s)" -ForegroundColor Yellow
@@ -136,4 +136,3 @@ This script assumes you have permissions to all sites in the tenant. We recommen
 
   </TabItem>
 </Tabs>
-
