+++ src/index.spec.ts
@@ -90,8 +90,8 @@ describe('Lazy loading commands', () => {
       'spo sp permissionrequest deny',
       'spo sp permissionrequest list',
       'spo sp set',
-      'spo web rule list',
-      'spo web rule remove',
+      'spo web alert list',
+      'spo web alert remove',
       'teams user add',
       'teams user list',
       'teams user remove',
